pyproject.toml
[project]
name = "litecache2"
version = "0.1.5"
description = "A lightweight async TCP dictionary server with RESP support"
readme = "README.md"
authors = [
    { name = "Joe Pasquantonio", email = "joe.pasquantonio@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "colorlog>=6.9.0",
    "pyyaml>=6.0.2",
    "uvloop>=0.21.0",
]

[project.scripts]
litecache = "litecache.server:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/litecache"]  # Directory containing your code

[tool.uv]
dev-dependencies = ["pytest>=7.0"]
